@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        v3.19.6
 // source: proto/user.api.proto
 
@@ -567,7 +567,7 @@ func file_proto_user_api_proto_rawDescGZIP() []byte {
 
 var file_proto_user_api_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_proto_user_api_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_proto_user_api_proto_goTypes = []interface{}{
+var file_proto_user_api_proto_goTypes = []any{
 	(UserField)(0),        // 0: UserField
 	(*User)(nil),          // 1: User
 	(*UserId)(nil),        // 2: UserId
@@ -607,7 +607,7 @@ func file_proto_user_api_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_proto_user_api_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_user_api_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*User); i {
 			case 0:
 				return &v.state
@@ -619,7 +619,7 @@ func file_proto_user_api_proto_init() {
 				return nil
 			}
 		}
-		file_proto_user_api_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_user_api_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*UserId); i {
 			case 0:
 				return &v.state
@@ -631,7 +631,7 @@ func file_proto_user_api_proto_init() {
 				return nil
 			}
 		}
-		file_proto_user_api_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_user_api_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateUserReq); i {
 			case 0:
 				return &v.state
@@ -643,7 +643,7 @@ func file_proto_user_api_proto_init() {
 				return nil
 			}
 		}
-		file_proto_user_api_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_user_api_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*ListUsersReq); i {
 			case 0:
 				return &v.state
@@ -655,7 +655,7 @@ func file_proto_user_api_proto_init() {
 				return nil
 			}
 		}
-		file_proto_user_api_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_user_api_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*UserFilter); i {
 			case 0:
 				return &v.state
@@ -667,7 +667,7 @@ func file_proto_user_api_proto_init() {
 				return nil
 			}
 		}
-		file_proto_user_api_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_user_api_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*ListUsersResp); i {
 			case 0:
 				return &v.state
